Create createpass
Createpass is a unique password generator. It can generate the following types of passwords: Single-handed passwords (can typed with ONLY the right or left hand on the keyboard), pseudorandom (using dev/urandom as the seed, with another option to pipe that output through base64), hexadecimal, and passwords using a unique date seed.