Create juliantime
Some years ago, I wrote a novella called "The Sleeper Race", which I published on deviantART (http://fav.me/d4srzzh). In that novella, I hypothesized a new way to how time is perceived and counted in my post-apocalyptic future world. Days were counted in Modified Julian Date with 5+ decimal places to indicate fractions of a day (hours, minutes, and seconds). This changes the day from a division of 24 hours into a division of tenths and hundredths. Each primary division of tenths (0.1) is approximately 2.4 hours long. Thus when the time is 0.25 it is 6am, 0.5 is noontime, and when it is 0.75 it is 6pm.
After working on this script for about 5 days total I am happy to present my "Modified Julian Date BASH Realtime Clock". Let me break down the name to explain it.

Modified Julian Date - Modified Julian Date is what this script converts Gregorian time (which most people consider standard) to.
Realtime - BASH has outputs for seconds and nanoseconds. I tested the calculation loop completely uncorked at one point, and it's accurate down to a nanosecond. The screen refresh and visual output likely drops it's efficiency a bit (I don't know how fast BASH can refresh a small term window) but it's basically realtime (give or take a nanosecond).
BASH - Written ONLY in BASH.
Clock - Because it computes the Modified Julian Date in realtime, as opposed to a website that must be manually reloaded or refreshed, this can be used as an actual and accurate form of timekeeping.