Create tradewind
I've been playing the game Tradewinds lately. Designed this to keep track of market prices, so I know when there's a good deal or not.

Having some trouble with this those... I want it to compare the current price with the low price and if current is lower than the listed low, then low is updated to the new low. Same with high prices; I want it to update as a higher price comes up.